finish vg-lite adapt

+++ src/draw/lv_draw_vector.h
@@ -108,7 +108,7 @@ typedef struct {
 typedef struct {
     lv_vector_gradient_style_t style;
     lv_gradient_stop_t   stops[LV_GRADIENT_MAX_STOPS];  /**< A gradient stop array */
-    uint8_t              stops_count;                   /**< The number of used stops in the array */
+    uint16_t             stops_count;                   /**< The number of used stops in the array */
     float x1;
     float y1;
     float x2;
